+++ auth/index.js
@@ -1,5 +1,4 @@
-const bcrypt = require('bcrypt');
-const cipher = require('../middlewares/hash');
+const cipher = require('../middlewares/cipher');
 const userModel = require('../database/models');
 const requestHelper = require('../helpers');
 
@@ -7,7 +6,6 @@ const authenticate = async (req, res, next) => {
   const { username, password } = req.body;
   try {
     const check = await userModel.findUser(username).first();
-    // const checkPassword = await bcrypt.compareSync(password, check.password);
     const checkPassword = await cipher.compare(password, check.password);
     if (check.username === username && checkPassword) {
       // eslint-disable-next-line require-atomic-updates
